backend/requirements.txt
# --- 核心框架 ---
fastapi
uvicorn
python-multipart

# --- 数据库与驱动 (修复点) ---
sqlalchemy
psycopg2-binary

# --- 安全与鉴权 ---
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]

# --- AI 服务 ---
openai>=1.0.0
python-dotenv